[FIX]account_analytic_report: Corrección en data de pyg res_partner_farm_data: Número de empleados de entero a decimal.

+++ project-addons/res_partner_farm_data/models/employee_count.py
@@ -31,7 +31,7 @@ class EmployeeFarmCount(custom_models.HistoricalModel):
     partner_id = fields.Many2one('res.partner', 'Partner', readonly=True)
     date = fields.Date('Date')
     user_id = fields.Many2one('res.users', 'User', readonly=True)
-    quantity = fields.Integer('Employees',
+    quantity = fields.Float('Employees',
                               states={'current': [('readonly', True)]})
     state = fields.Selection(
         (('current', 'Current'), ('history', 'History')), 'State',
